requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
gspread
playwright
pystray
pillow
plyer
